feat: added error handling and validations

+++ src/main/java/pt/ridenexus/vehicle/persistence/model/VehicleEntity.java
@@ -47,7 +47,7 @@ public class VehicleEntity extends BaseEntity {
     private LocalDate licensePlateDate;
 
     @Column(name = "weight")
-    private Double weight;
+    private Integer weight;
 
     @Column(name = "mileage")
     private Integer mileage;
